Allow inline references for sample constituents
