+++ pydatalab/pydatalab/models/entries.py
@@ -6,10 +6,9 @@ from pydantic import BaseModel, Field, root_validator
 from pydatalab.models.relationships import TypedRelationship
 from pydatalab.models.utils import (
     JSON_ENCODERS,
-    HumanReadableIdentifier,
+    EntryReference,
     IsoformatDateTime,
     PyObjectId,
-    Refcode,
 )
 
 
@@ -74,37 +73,3 @@ class Entry(BaseModel, abc.ABC):
         allow_population_by_field_name = True
         json_encoders = JSON_ENCODERS
         extra = "ignore"
-
-
-class EntryReference(BaseModel):
-    """A reference to a database entry by ID and type.
-
-    Can include additional arbitarary metadata useful for
-    inlining the item data.
-
-    """
-
-    type: str
-    immutable_id: Optional[PyObjectId]
-    item_id: Optional[HumanReadableIdentifier]
-    refcode: Optional[Refcode]
-
-    @root_validator
-    def check_id_fields(cls, values):
-        """Check that only one of the possible identifier fields is provided."""
-        id_fields = ("immutable_id", "item_id", "refcode")
-
-        # Temporarily remove refcodes from the list of fields to check
-        # until it is fully implemented
-        if values.get("refcode") is not None:
-            values["refcode"] = None
-        if all(values.get(f) is None for f in id_fields):
-            raise ValueError(f"Must provide at least one of {id_fields!r}")
-
-        if sum(1 for f in id_fields if values.get(f) is not None) > 1:
-            raise ValueError("Must provide only one of {id_fields!r}")
-
-        return values
-
-    class Config:
-        extra = "allow"
